crop-disease-backend/requirements.txt
flask
flask-cors
gunicorn
tensorflow
torch
transformers
pillow
